Making DataFile indexing optional in search and making autocomplete optional and turned off by default as its implementation is incomplete
