+++ tardis/search/datafile_index.py
@@ -1,0 +1,59 @@
+# -*- coding: utf-8 -*-
+
+#
+# Copyright (c) 2010-2011, Monash e-Research Centre
+#   (Monash University, Australia)
+# Copyright (c) 2010-2011, VeRSI Consortium
+#   (Victorian eResearch Strategic Initiative, Australia)
+# All rights reserved.
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+#    *  Redistributions of source code must retain the above copyright
+#       notice, this list of conditions and the following disclaimer.
+#    *  Redistributions in binary form must reproduce the above copyright
+#       notice, this list of conditions and the following disclaimer in the
+#       documentation and/or other materials provided with the distribution.
+#    *  Neither the name of the VeRSI, the VeRSI Consortium members, nor the
+#       names of its contributors may be used to endorse or promote products
+#       derived from this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE REGENTS AND CONTRIBUTORS ``AS IS'' AND ANY
+# EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE REGENTS AND CONTRIBUTORS BE LIABLE FOR ANY
+# DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+# (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+# LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+# ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+# SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
+
+'''
+search indexes for single search
+
+.. moduleauthor:: Shaun O'Keefe  <shaun.okeefe@versi.edu.au>
+.. moduleauthor:: James Wettenhall <james.wettenhall@monash.edu>
+
+'''
+import os
+from haystack import indexes
+
+from tardis.tardis_portal.models import DataFile
+
+
+class DataFileIndex(indexes.SearchIndex, indexes.Indexable):
+    text = indexes.CharField(document=True)
+    datafile_filename = indexes.CharField(model_attr='filename')
+    experiment_id_stored = indexes.MultiValueField(indexed=True, stored=True)
+    dataset_id_stored = indexes.IntegerField(model_attr='dataset__id')
+
+    def prepare_text(self, obj):
+        return os.path.join(obj.directory or '', obj.filename)
+
+    def prepare_experiment_id_stored(self, obj):
+        return [exp.id for exp in obj.dataset.experiments.all()]
+
+    def get_model(self):
+        return DataFile
